Fix bug in getting container gateway. ContainerInfo.NetworkSettings.Gateway is not available when a user defined network is used.
However, it is availble in ContainerInfo.NetworkSettings.Networks[my_network].Gateway.

The impact of having it unavailable is that there is there is a downsteam bug that makes dockerHostIp becomes null, instead of returning the default of localhost.

This causes other downstreams that does not expect null from this method to throw exceptions